requirements.txt
matplotlib
pandas
librosa
numpy
torch 
torchvision 
torchaudio
IPython
pyyaml
tqdm